Factor out politician page components
