+++ components/PoliticianPage/ElectionInfo/ElectionInfoSection.module.scss
