+++ components/PoliticianPage/ElectionInfo/ElectionInfoSection.tsx
