+++ components/PoliticianPage/EndorsementsSection/EndorsementsSection.module.scss
@@ -1,0 +1,1 @@
+@import "../../../pages/politicians/PoliticianPage.module.scss";
