+++ components/index.tsx
@@ -5,12 +5,12 @@ import { BetaNotice } from "./BetaNotice/BetaNotice";
 import { BillCard } from "./BillCard/BillCard";
 import { Button } from "./Button/Button";
 import { EditVotingGuideCandidate } from "./Ballot/Race";
-import { ElectionInfoSection } from "./Politician/ElectionInfo/ElectionInfoSection";
+import { ElectionInfoSection } from "./PoliticianPage/ElectionInfo/ElectionInfoSection";
 import { ElectionSelector } from "./ElectionSelector/ElectionSelector";
 import { FieldSet } from "./FieldSet/FieldSet";
 import { FlagSection, FlagColor } from "./FlagSection/FlagSection";
 import { Footer } from "./Footer/Footer";
-import { HeaderSection } from "./Politician/HeaderSection/HeaderSection";
+import { HeaderSection } from "./PoliticianPage/HeaderSection/HeaderSection";
 import { HomePageButton } from "./HomePageButton/HomePageButton";
 import { ImageWithFallback } from "./ImageWithFallback/ImageWithFallback";
 import { Layout } from "./Layout/Layout";
